requirements.txt
Flask>=3.1.0
Flask-SQLAlchemy>=3.1.1
transformers>=4.48.3
langchain>=0.3.17
gTTS>=2.5.4
Pillow>=11.1.0
requests>=2.32.3
